chore: split sorting e2e specs into per-domain files
Replace the consolidated sorting.test.ts with per-domain specs: restore the accounts/assets/transactions specs and add securities and portfolio specs for the newly sortable tables. Keep the transactions server-sort refetch-race fix.
